JENKINS-60940: Migrate translated tests from GitAPITestCase to GitAPICaseTest class
